// packages/shared/src/enums.ts
export enum UserRole {
  ADMIN = 'ADMIN',
  OFFICER = 'OFFICER',
  ANALYST = 'ANALYST',
  REVIEWER = 'REVIEWER',
  MP = 'MP',
  CONTRACTOR = 'CONTRACTOR',
  CITIZEN = 'CITIZEN',
  FIELD_OFFICER = 'FIELD_OFFICER',
  VIEWER = 'VIEWER'
}